clean up non-parameterized tests in the wrong place
